Make shared source dir for 2.11 and 2.12
