+++ main/compile.sbt
@@ -1,0 +1,10 @@
+Compile / unmanagedSourceDirectories ++= {
+  val sharedSourceDir = (ThisBuild / baseDirectory).value / "main/src/main"
+  println(s"sharedSourceDir $sharedSourceDir")
+  val additionalDirs = CrossVersion.partialVersion(scalaVersion.value) match {
+    case Some((2, 11) | (2, 12)) => Seq(sharedSourceDir / "scala-2.11_2.12")
+    case _ => Seq.empty
+  }
+  println(s"additionalDirs: $additionalDirs")
+  additionalDirs
+}
