+++ main/src/main/scala-2.11_2.12/org/clulab/scala/WrappedBuffer.scala
