+++ main/src/main/scala-2.11_2.12/org/clulab/struct/DependencyMap.scala
@@ -6,10 +6,7 @@ class DependencyMap protected extends mutable.HashMap[Int, DirectedGraph[String]
   override def initialSize: Int = 2 // we have very few dependency types, so let's create a small hash to save memory
 }
 
-object DependencyMap {
-  val STANFORD_BASIC: Int = DependencyMapNames.STANFORD_BASIC
-  val STANFORD_COLLAPSED: Int = DependencyMapNames.STANFORD_COLLAPSED
-  val SEMANTIC_ROLES: Int = DependencyMapNames.SEMANTIC_ROLES
+object DependencyMap extends DependencyMapNames {
 
   def apply(): DependencyMap = new DependencyMap()
 }
