--- main/src/main/scala-2.12/org/clulab/scala/WrappedArrayBuffer.scala
@@ -1,3 +1,0 @@
-package org.clulab.scala
-
-object WrappedArrayBuffer
