+++ main/src/main/scala-2.13/org/clulab/struct/DependencyMap.scala
@@ -2,11 +2,7 @@ package org.clulab.struct
 
 import scala.collection.mutable
 
-object DependencyMap {
-  val STANFORD_BASIC: Int = DependencyMapNames.STANFORD_BASIC
-  val STANFORD_COLLAPSED: Int = DependencyMapNames.STANFORD_COLLAPSED
-  val SEMANTIC_ROLES: Int = DependencyMapNames.SEMANTIC_ROLES
-
+object DependencyMap extends DependencyMapNames {
   // This was previously a class inheriting from HashMap.  However,
   // [warn] ...: inheritance from class HashMap in package mutable is deprecated (since 2.13.0): HashMap will be made final; use .withDefault for the common use case of computing a default value.
   type DependencyMap = mutable.HashMap[String, DirectedGraph[String]]
