+++ main/src/main/scala-2.13/org/clulab/utils/NamedFuture
@@ -1,0 +1,19 @@
+
+  object NamedFuture {
+
+    def apply[T](name: String)(body: => T)(implicit @deprecatedName("execctx") executor: ExecutionContext): Future[T] = {
+      Future {
+        val thread = Thread.currentThread
+        val oldName = thread.getName
+        val result = try {
+          thread.setName(s"$oldName ($name)")
+          body
+        }
+        finally {
+          thread.setName(oldName)
+        }
+
+        result
+      }(executor)
+    }
+  }
