+++ main/src/main/scala-2.13/org/clulab/utils/ThreadUtils.scala
@@ -7,7 +7,6 @@ import scala.collection.parallel.ForkJoinTaskSupport
 import scala.collection.parallel.ParIterable
 import scala.collection.parallel.ParSeq
 import scala.collection.parallel.ParSet
-import scala.concurrent.{ExecutionContext, Future}
 
 object ThreadUtils {
 
@@ -37,23 +36,4 @@ object ThreadUtils {
   }
 
   def parallelize[T](set: Set[T]): ParSet[T] = set.par
-
-  object NamedFuture {
-
-    def apply[T](name: String)(body: => T)(implicit @deprecatedName("execctx") executor: ExecutionContext): Future[T] = {
-      Future {
-        val thread = Thread.currentThread
-        val oldName = thread.getName
-        val result = try {
-          thread.setName(s"$oldName ($name)")
-          body
-        }
-        finally {
-          thread.setName(oldName)
-        }
-
-        result
-      }(executor)
-    }
-  }
 }
