+++ main/src/main/scala/org/clulab/struct/DependencyMapNames.scala
@@ -1,6 +1,6 @@
 package org.clulab.struct
 
-object DependencyMapNames {
+trait DependencyMapNames {
   val STANFORD_BASIC = 0 // basic Stanford dependencies
   val STANFORD_COLLAPSED = 1 // collapsed Stanford dependencies
   val SEMANTIC_ROLES = 2 // semantic roles from CoNLL 2008-09, which includes PropBank and NomBank
